+++ src/Lib/Config/Description/InitializableDescriptionInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace PP\Lib\Config\Description;
+
+interface InitializableDescriptionInterface {
+
+	/**
+	 * Called when PXApplication wakes up from cache
+	 *
+	 * @param \PXApplication $app
+	 * @return void
+	 */
+	public function initialize(\PXApplication $app);
+}